Run compile and test in CI for all scala versions
The project specifies a number of scala versions in our build.sbt, but
before now CI was running `compile` and `test`, which only use the
current scala version. This meant it was possible for CI to pass if the
code compiled successfully in the current version but failed in one of
the versions in `crossScalaVersions`.

This commit updates the CI script to use `+compile` and `+test`, which
run those tasks for all versions in `crossScalaVersions`.