Fix version-number confusion - skip straight to v1.0.0
This library has had some version -number confusion:

* 0.4.4 - 2023/03/17 - updated version.sbt
* 0.5.0 - 2023/04/28 - did not update version.sbt - https://github.com/guardian/fastly-api-client/releases/tag/v0.5.0
* 0.6.0 - 2023/05/05 - did not update version.sbt - did not get a `git` tag
* 0.4.5 - 2024/07/26 - picked up this smaller version number based on what was in version.sbt
* 0.4.6 - 2023/07/30
* ...etc

This repo adopted `gha-scala-library-release-workflow` with https://github.com/guardian/fastly-api-client/pull/47
in 2024/07/26 - before that the release procedure was much more manual and error-prone!

To get past the confusion, it's best to jump to a version number higher than 0.6.0 -
and as this library, like all libraries that use `gha-scala-library-release-workflow`,
is using `early-semver`, it makes sense to move past 0.y.z as a version number - under
`early-semver`:

> When the major version is 0, a minor version increment MAY contain both source and binary breakages, but a patch version increment MUST remain binary compatible.

To avoid the complication of trying to understand that, better to just get to v1 and above.